requirements.txt
pynacl
discord.py
dotenv
asyncio
